Banner accessibility improvements
